# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nika Web Services",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/packages": {
            "get": {
                "tags": [
                    "پکیج"
                ],
                "summary": "لیست پکیج ها بصورت صفحه بندی",
                "description": "",
                "operationId": "753372c28cb73b97aee344102ef4a873",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "شماره صفحه",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "تعداد نمایش در هر صفحه",
                        "schema": {
                            "type": "number"
                        },
                        "example": 10
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "جستجوی عنوان",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "پکیج"
                ],
                "summary": "ثبت پکیج",
                "description": "",
                "operationId": "f7453aeadb2094b208e4e073c9207210",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "age",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "عنوان",
                                        "type": "string"
                                    },
                                    "age": {
                                        "description": "سن",
                                        "type": "number"
                                    },
                                    "price": {
                                        "description": "قیمت",
                                        "type": "number"
                                    },
                                    "is_completed": {
                                        "description": "وضعیت تکمیل - بصورت boolean باید باشد",
                                        "type": "string",
                                        "enum": [
                                            1,
                                            0
                                        ]
                                    },
                                    "video": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "description": "وضعیت Active:فعال Inactive:غیرفعال",
                                        "type": "string",
                                        "enum": [
                                            "Active",
                                            "Inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ثبت با موفقیت انجام شد",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/packages/{id}": {
            "get": {
                "tags": [
                    "پکیج"
                ],
                "summary": "نمایش اطلاعات پکیج",
                "description": "",
                "operationId": "fb1e928dec538b75507809f04938790c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه پکیج",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "پکیج"
                ],
                "summary": "بروزرسانی پکیج",
                "description": "",
                "operationId": "336f10131c4c998df62d5ae774ea6740",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه پکیج",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "title",
                                    "age",
                                    "price"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "این مقدار باید بصورت ثابت شود",
                                        "type": "string",
                                        "default": "put",
                                        "enum": [
                                            "put"
                                        ]
                                    },
                                    "title": {
                                        "description": "عنوان",
                                        "type": "string"
                                    },
                                    "age": {
                                        "description": "سن",
                                        "type": "number"
                                    },
                                    "price": {
                                        "description": "قیمت",
                                        "type": "number"
                                    },
                                    "is_completed": {
                                        "description": "وضعیت تکمیل - بصورت boolean باید باشد",
                                        "type": "string",
                                        "enum": [
                                            1,
                                            0
                                        ]
                                    },
                                    "status": {
                                        "description": "وضعیت Active:فعال Inactive:غیرفعال",
                                        "type": "string",
                                        "enum": [
                                            "Active",
                                            "Inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "پکیج"
                ],
                "summary": "حذف پکیج",
                "description": "",
                "operationId": "c533535e347ef0e2d5756f6079db950c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه پکیج",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "توکن را به همراه Bearer وارد کنید.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}